requirements.txt
simplejson~=3.19.3
requests~=2.32.3
python-dotenv~=1.0.1